requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Real-time Communication
python-socketio==5.10.0

# File Processing
pillow==10.1.0
python-multipart==0.0.6

# Content Processing
markdown==3.5.1
bleach==6.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2

# Logging
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0



